+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+
+### Todo
+
+* on(…) events with enums
+
+### 2.0 (upcoming)
+
+* Add new DigitalInputOverridable class which allows to manually override the input value
+  that is returned (also affects emitted events).
+
 ### 1.2
 
 * **breaking** – `DigitalInput` signature changed, the input mode can now be chosen